Extract TreeLeafCulling class from TreeRenderer
Refactor TreeRenderer to separate compute culling pipelines into a
dedicated TreeLeafCulling class. This reduces TreeRenderer from ~2180
lines to ~680 lines and uses DescriptorManager utilities (LayoutBuilder,
SetWriter) throughout to reduce boilerplate code.

- Create TreeLeafCulling.h/cpp with all culling pipelines and buffers
- Simplify TreeRenderer to delegate culling to TreeLeafCulling
- Use DescriptorManager::LayoutBuilder for descriptor set layouts
- Use DescriptorManager::SetWriter for descriptor set updates
